requirements.txt
httpx
quart
quart_cors
python-dotenv